Add files via upload
merging of windows and linux python script into one

+++ sourcewindows/SuperPatcherGSI.py
@@ -2,25 +2,30 @@ import os
 import argparse
 import shutil
 import pathlib
+import platform
 
-TempDIR = os.getcwd() + "\\" + ".temp"
+
+platform = 1 if platform.system() == 'Linux' else 0 if platform.system() == 'Windows' else 2
+TempDIR = (os.getcwd() + "/" + ".temp") if platform == 1 else (os.getcwd() + "\\" + ".temp") if platform == 0 else None
 HERE = os.path.realpath(os.path.dirname(__file__))
 
-def dir_path(string):
-    if os.path.isdir(string):
-        return os.path.realpath(string)
-    else:
-        raise NotADirectoryError(string)
 
-parser = argparse.ArgumentParser(add_help=False)
-parser.add_argument('-i' , '--input' , type=argparse.FileType('r') , help='''input the super.img that is going to be modifed, if super.img is sparse its going to temporarily be unsparsed''')
+def mod_path(string):
+    if os.path.isfile(string):
+        return string
+    elif os.path.isdir(string):
+        return os.path.realpath(string)
+
+parser = argparse.ArgumentParser(add_help=True)
+parser.add_argument('-i' , '--input' , type=mod_path , help='''Input the super.img that is going to be modifed if super.img is sparse its going to temporarily be unsparsed, you can also input a directory with files to be packed to an super.img''')
 parser.add_argument('-o' , '--output', help="Directs the output to a name of your choice")
 parser.add_argument('-s' , '--SLOT' , type=int , help="number of slots on the device can only be 1 (A) or 2 (A/B)")
-parser.add_argument('-p', '--path' , type=dir_path , help="specifies input to be a folder and not an super.img")
 args = parser.parse_args()
+
 
 # err check
 def check():
+    
     err = ""
     try:
         if args.SLOT == 1 or args.SLOT == 2:
@@ -28,21 +33,17 @@ def check():
         else:
             print("Invalid Slot number ({slot})".format(slot=args.SLOT))
             err += " &SLOT"
-        if args.input == None:
-            if args.path != "":
-                pass
-            else:
-                print("Invalid path at PATH please input a valid directory")
-                err += " &InvalidPath"
-        elif args.input.name.endswith(".img"):
+        if args.input.endswith(".img") and os.path.isfile(args.input):
             pass
+        elif os.path.isdir(args.input):
+            pass 
         else:
-            print("Invalid Format at INPUT please use .img file")
+            print("Invalid Format at INPUT please use .img file or valid directory")
             err += " &InvalidFormatINPUT"
         if args.output.endswith(".img"):
             pass
         else:
-            print("Invalid Format at INPUT please use .img file")
+            print("Invalid Format at OUTPUT please use .img file")
             err += " &InvalidFormatOUTPUT"
         if err == "":
             err = "OK"
@@ -52,13 +53,17 @@ def check():
         err = "Flag AttributeError"
     return err
 
-# unpack / replacing
 
-def lpunpack(): # PAIN just PAIN
+# unpack / replacing
+def lpunpack():
     #!/usr/bin/python  # -*- coding: latin-1 -*-
-    os.system("powershell {command}"
+    if platform == 1:
+        os.system("python3 '{dir}/lpunpack.py' {superimg} '{tempdir}'".format(superimg=args.input, tempdir=TempDIR , dir=HERE))
+    if platform == 0:
+        os.system("powershell {command}"
         .format(command=repr(".\'{dir}\\lpunpack.exe' {superimg} '{tempdir}'"
-            .format(superimg=args.input.name, tempdir=TempDIR , dir=HERE))))
+            .format(superimg=args.input, tempdir=TempDIR , dir=HERE))))
+
 
 def IMGmanipulation(): # choose an img file to be replaced
     TempImgList = os.listdir(TempDIR)
@@ -66,7 +71,6 @@ def IMGmanipulation(): # choose an img file to be replaced
     for img in TempImgList:
             if img.endswith(".img"):
                 d += 1
-
     print("Chosse Operation:")
     if d != 0:
         print("1. Delete Partition")
@@ -81,7 +85,7 @@ def IMGmanipulation(): # choose an img file to be replaced
     if Operation == "2" or Operation == "1":
         for img in TempImgList:
             if img.endswith(".img"):
-                print("option number " + str(i) + " " + TempImgList[i] + " size of (" + str(os.path.getsize(TempDIR + "\\" + img)) + ") bytes")
+                print("option number " + str(i) + " " + TempImgList[i] + " size of (" + str(os.path.getsize(TempDIR + "/" + img)) + ") bytes")
             i += 1
     while (True):
         try:
@@ -92,7 +96,10 @@ def IMGmanipulation(): # choose an img file to be replaced
                 while (True):
                     try:
                         if Operation == "1": # just making sure
-                            os.remove(TempDIR + "\\" + TempImgList[int(imgnum)])
+                            if platform == 1:
+                                os.remove(TempDIR + "/" + TempImgList[int(imgnum)])
+                            if platform == 0:
+                                os.remove(TempDIR + "\\" + TempImgList[int(imgnum)])
                             print("Partition Deleted!")
                             break
                         elif Operation == "2":
@@ -103,7 +110,10 @@ def IMGmanipulation(): # choose an img file to be replaced
                                     if replacmentpath.endswith(".img"):
                                         redo = input("Are you sure this is the path to file (Y/n): ")
                                         if redo == "Y" or redo == "y" or redo == "yes" or redo == "Yes" or redo == "": # just making sure
-                                            shutil.copy(replacmentpath , TempDIR + "\\" + TempImgList[int(imgnum)])
+                                            if platform == 1:
+                                                shutil.copy(replacmentpath , TempDIR + "/" + TempImgList[int(imgnum)])
+                                            if platform == 0:
+                                                shutil.copy(replacmentpath , TempDIR + "\\" + TempImgList[int(imgnum)])
                                             print("Partition Replaced!")
                                             break
                                     elif replacmentpath.endswith(" "):  # sometimes people tap space it ruins it and it can and will get confusing
@@ -131,14 +141,17 @@ def IMGmanipulation(): # choose an img file to be replaced
                                     name = ""
                                 except ValueError:
                                     print("Please Put a Valid Number!")
-                            os.system("powershell {command}"
+                            if platform == 1:
+                                os.system("dd if=/dev/zero of='{tempdir}/{name}.img' bs=1 count={size}".format(tempdir=TempDIR , name=name , size=size))
+                            if platform == 0:
+                                os.system("powershell {command}"
                                 .format(command="fsutil file createnew '{tempdir}\\{name}.img' {size}"
                                     .format(tempdir=TempDIR , name=name , size=int(size))))
                             break
                     except ValueError:
                         print("Please Put a Valid Number!")
                 redo = input("Replace/Delete/Add another (Y/n): ")
-                if redo == "Y" or redo == "y" or redo == "yes" or redo == "Yes": # just making sure
+                if redo == "Y" or redo == "y" or redo == "yes" or redo == "Yes" or redo == "": # just making sure
                     IMGmanipulation()
                 break
             else:
@@ -146,17 +159,6 @@ def IMGmanipulation(): # choose an img file to be replaced
         except ValueError:
             print("Please Put a Number In!")
 
-def IMGsizeCALC(): # calculate size
-    totalsize = 5120000 # a bit of overhead
-    TempImgList = os.listdir(TempDIR)
-    i = 0
-    for img in TempImgList:
-        if img.endswith(".img"):
-            totalsize += os.path.getsize(TempDIR + "\\" + img)
-        i += 1
-    reminder = totalsize % 512 # making devision by 512
-    totalsize += reminder
-    return totalsize
 
 # lpmake
 def lpmake(devicesize , metadatasize):
@@ -164,19 +166,35 @@ def lpmake(devicesize , metadatasize):
     sparse = input("make sparse (flashable with fastboot) ? (Y/n): ")
     if sparse == "Y" or sparse == "y" or sparse == "yes" or sparse == "Yes" or sparse == "": # just making sure
         lpmake_args += " --sparse"
-    
-    lpmake_args = lpmake_add_args(lpmake_args)
+    if platform == 1:
+        lpmake_args = lpmake_add_args_linux(lpmake_args)
+    if platform == 0:
+        lpmake_args = lpmake_add_args_windows(lpmake_args)
     print("============================")
     print("    using these flags:")
     print("============================")
     print(lpmake_args)
     print("============================")
-    
-    return os.system("powershell {command}"
-        .format(command=".\'{dir}\\lpmake.exe' {lpargs}"
+
+    if platform == 1:
+        return os.system("'{dir}/bin/lpmake' {lpargs}".format(lpargs=lpmake_args , dir=HERE))
+    if platform == 0:
+        return os.system("powershell {command}"
+        .format(command=".\'{dir}\\bin\\lpmake.exe' {lpargs}"
             .format(lpargs=lpmake_args , dir=HERE)))
 
-def lpmake_add_args(lpmake_args):
+
+def lpmake_add_args_linux(lpmake_args):
+    TempImgList = os.listdir(TempDIR)
+    for img in TempImgList:
+        if img.endswith(".img"):
+            lpmake_args += " --partition={name}:none:{size}".format(name=os.path.splitext(img)[0] , size=os.path.getsize(TempDIR + "/" + img))
+            if os.path.getsize(TempDIR + "/" + img) != 0:
+                lpmake_args += " --image={name}={filedir}".format(name=os.path.splitext(img)[0] , filedir=(TempDIR + "/" + img))
+    return lpmake_args
+
+
+def lpmake_add_args_windows(lpmake_args):
     TempImgList = os.listdir(TempDIR)
     for img in TempImgList:
         if img.endswith(".img"):
@@ -185,19 +203,41 @@ def lpmake_add_args(lpmake_args):
                 lpmake_args += " --image={name}={filedir}".format(name=os.path.splitext(img)[0] , filedir=repr((TempDIR + "\\" + img)))
     return lpmake_args
 
-def testdvi512(num):
-    if num % 512 == 0:
-        return True
-    else:
-        return False
+
+def IMGsizeCALC(): # calculate size
+    totalsize = 5120000 # a bit of overhead
+    TempImgList = os.listdir(TempDIR)
+    i = 0
+    for img in TempImgList:
+        if img.endswith(".img"):
+            if platform == 1:
+                totalsize += os.path.getsize(TempDIR + "/" + img)
+            if platform == 0:
+                totalsize += os.path.getsize(TempDIR + "\\" + img)
+        i += 1
+    reminder = totalsize % 512 # making devision by 512
+    totalsize += reminder
+    return totalsize
+
 
 def copytotemp(cloneto , clonefrom):
     files = os.listdir(clonefrom)
     os.makedirs(cloneto)
     for file in files:
         print("copying " + file)
-        shutil.copy2(clonefrom + "\\" + file , cloneto)
+        if platform == 1:
+            shutil.copy2(clonefrom + "/" + file , cloneto)
+        if platform == 0:
+            shutil.copy2(clonefrom + "\\" + file , cloneto)
     #shutil.copytree(clonefrom, cloneto)
+
+
+def testdvi512(num):
+    if num % 512 == 0:
+        return True
+    else:
+        return False
+
 
 def main():
     err = check()
@@ -207,12 +247,12 @@ def main():
         return err
     else:
         print("flags successfully verified and appear to be correct, error code ({error})".format(error=err))
-    if args.path != None:
+    if os.path.isdir(args.input):
         print("============================")
         print("   copying to temp dir...")
         print("============================")
-        copytotemp(TempDIR , args.path)
-    else:
+        copytotemp(TempDIR , args.input)
+    elif args.input.endswith(".img"):
         print("============================")
         print("        unpacking...")
         print("============================")
@@ -221,8 +261,9 @@ def main():
     print("      img manipulation ")
     print("============================")
     
-    
-    IMGmanipulation() # IMG manipulation of selected partition
+    do = input("Replace/Delete/Add (Y/n): ")
+    if do == "Y" or do == "y" or do == "yes" or do == "Yes" or do == "": # just making sure
+        IMGmanipulation() # IMG manipulation of selected partition
     #let user choose size
     metadatasize = 512000
     devicesize = IMGsizeCALC()
